Gramma/typo
there was a word missing on line 29. 

I really like the sound of this project and will mull some of the ideas over to see if I can come up with something more constructive than a simple correction.  Totally agree with the concept and would like to help move if forward.