Fix one source of config disconnect problem
The MainActivity is using a fragment stack on a container for About and
Config but a different stack on a different container for
NearbyDevices. The result is a missing onPause call when entering
BeaconConfigFragment. This kept the BLE scan running at LOW_LATENCY.
Scanning and GATT connections cannot occur at the same time so this was
causing a problem.