Remove jitter
Increase the "lost" timeout from 5 seconds to 120 seconds to reduce jitter. Entries are removed from the list after this number of seconds. 

The far region is now alphabetically sorted so beacons like "About" sorts high even though it is far away... and if it isn't heard in 1 or 2 cycles (5 seconds) it disappears from the list and causes visual jitter.