slow metadata response fix
After scan timeout of 3 seconds, list items that do not have metadata
are shown with just the advertised beacon url. As the metadata is
received (after the 3 seconds), the correlate list item ui is updated
to show that metadata. Also, view recycling is now handled so that list
items are always cleared in getView() before updating (since a recycled
view may contain another url’s metadata from a previous cycle)