notification ux improvements
Currently if two or more beacons exist, two notifications are created
plus a summary notification. The creation order was backwards so that
the two notifications would be created and then the summary. This made
two notification appear and then collapse into the summary. Here we
make the summary first, so the two notifications are not shown first.