Proxy Service includes "rank" value to better sort on clients
Currently the rank is set to the signal path loss, but may include other
signals in the future.  If the requests includes multiple values, they
will be sorted on the server automatically.  However, this rank value is
useful for merging separate replies from multiple independant metadata
requests on the client.