Putting logo into Readme
Due to the security constraints that come from stripping all CSS from a markdown file, formatting is VERY limited. I'd rather have it in the title but I just couldn't get it to line up properly. This layout is a bit odd, but it's the best I can come up with. If anyone else wants to take a crack, go for it....