[android] Use the DiscoveryService in the fragment
Instead of doing scanning independently in these two different classes,
we connect the NearbyBeaconFragment to the PwoDiscoveryService and
request PWOs through it.

This additionally means that we don't need to worry about coordinating
two scanners in the MainActivity.

What this change does not do is persist PWOs beyond the 10 second run
time of the discovery service.  That may be coming in a future change.