Allow Fatbeacon to serve large files
This change allows the Fatbeacon to serve large files, as opposed to
one message that is less than the mtu size. The central will call
the same characteristic until it receives a packet size smaller
than the mtu. The peripheral will slice the message and serve the
next slice when the central askes for it.