Allow for Changing the Connection Interval
This change allows us to change the connection interval through
system calls when we obtain the connection handle. It's hacky,
but it accomplishes the task. It's important to note a few things.
First, the Bluetooth specs require that a connection interval as low
as 7.5 ms. Second, there is a 1.25 multiplier on the min & max
settings.