Increase ci (#889)
* Change MTU size of BLE transfer packets

This change allows us to monkeypatch the increase in MTU size to
optimize the data transfer rate. Currently, we are setting it at
505, which is the limit the android can accept.

* Allow for Changing the Connection Interval

This change allows us to change the connection interval through
system calls when we obtain the connection handle. It's hacky,
but it accomplishes the task. It's important to note a few things.
First, the Bluetooth specs require that a connection interval as low
as 7.5 ms. Second, there is a 1.25 multiplier on the min & max
settings.